Restructing DCG to work in all directions and in partial mode.
The original design was designed in a such a way that it only worked in
the forward mode. Partial modes were also not suported. This is also
much more efficient than the original design as well. The main drawback
here is that representations are not directly in integer and string
format. Structure digit and word lists are required for this design.
Nonetheless, abstractions can be written atop this new structure to
remedy said caveats.